+++ frontend/app/index.tsx
@@ -1,351 +1,113 @@
 import React, { useState } from 'react';
+import { useRouter } from 'expo-router';
 
-interface SurveyState {
-  fever: boolean | null;
-  urgentSymptoms: string[];
-  severePain: boolean | null;
-  infectionSymptoms: string[];
-  recentFall: boolean | null;
-  canBearWeight: boolean | null;
-  chronicCondition: boolean | null;
-  worseSymptomsToday: boolean | null;
-  mentalHealthConcerns: string[];
-  dailyAssistance: string[];
-  medicationManagement: boolean | null;
-  poorNutrition: boolean | null;
-  loneliness: boolean | null;
-}
 
-const MedicalSurvey = () => {
-  const [state, setState] = useState<SurveyState>({
-    fever: null,
-    urgentSymptoms: [],
-    severePain: null,
-    infectionSymptoms: [],
-    recentFall: null,
-    canBearWeight: null,
-    chronicCondition: null,
-    worseSymptomsToday: null,
-    mentalHealthConcerns: [],
-    dailyAssistance: [],
-    medicationManagement: null,
-    poorNutrition: null,
-    loneliness: null,
-  });
+const HomePage = () => {
+  const router = useRouter();
+  // Track hover state for button interactions
+  const [isHovered, setIsHovered] = useState(false);
 
-  const [submitted, setSubmitted] = useState(false);
 
-  const containerStyle = {
-    maxWidth: '800px',
-    margin: '0 auto',
-    fontFamily: 'Arial, sans-serif',
-    height: '100vh',
-    display: 'flex',
-    flexDirection: 'column' as const,
+  const goToSurvey = () => {
+    router.push('/survey');
   };
 
-  const headerStyle = {
-    padding: '20px',
-    background: `linear-gradient(135deg, #d5002a 0%, #d3058b 100%)`,
-    color: 'white',
-    position: 'sticky' as const,
-    top: 0,
-    zIndex: 1,
-  };
-
-  const emergencyWarningStyle = {
-    backgroundColor: '#d5002a',
-    color: 'white',
-    padding: '20px',
-    marginBottom: '20px',
-    borderRadius: '8px',
-    boxShadow: '0 4px 6px rgba(213, 0, 42, 0.2)',
-  };
-
-  const warningHeaderStyle = {
-    display: 'flex',
-    alignItems: 'center',
-    gap: '10px',
-    marginBottom: '15px',
-    fontSize: '20px',
-    fontWeight: 'bold',
-  };
-
-  const emergencyListStyle = {
-    margin: '0',
-    padding: '0 0 0 20px',
-    listStyle: 'none',
-  };
-
-  const emergencyItemStyle = {
-    marginBottom: '10px',
-    display: 'flex',
-    alignItems: 'flex-start',
-    gap: '10px',
-  };
-
-  const formContainerStyle = {
-    flex: 1,
-    overflow: 'auto',
-    padding: '20px',
-    backgroundColor: '#f8f8f8',
-  };
-
-  const sectionStyle = {
-    marginBottom: '30px',
-    padding: '20px',
-    border: '1px solid rgba(213, 0, 42, 0.1)',
-    borderRadius: '8px',
-    backgroundColor: 'white',
-    boxShadow: '0 2px 4px rgba(0, 0, 0, 0.05)',
-  };
-
-  const questionStyle = {
-    marginBottom: '20px',
-  };
-
-  const labelStyle = {
-    display: 'block',
-    marginBottom: '10px',
-    fontWeight: 'bold',
-    color: '#333',
-  };
-
-  const buttonStyle = {
-    padding: '12px 24px',
-    background: `linear-gradient(135deg, #d5002a 0%, #d3058b 100%)`,
-    color: 'white',
-    border: 'none',
-    borderRadius: '4px',
-    cursor: 'pointer',
-    width: '100%',
-    fontSize: '16px',
-    transition: 'opacity 0.2s',
-    ':hover': {
-      opacity: 0.9,
-    },
-  };
-
-  const footerStyle = {
-    padding: '20px',
-    background: 'white',
-    borderTop: '1px solid rgba(213, 0, 42, 0.1)',
-    position: 'sticky' as const,
-    bottom: 0,
-  };
-
-  const checkboxContainerStyle = {
-    marginLeft: '20px',
-    marginBottom: '8px',
-  };
-
-  const radioContainerStyle = {
-    display: 'flex',
-    gap: '20px',
-    marginLeft: '20px',
-  };
-
-  const needsUrgentCare = () => {
-    return (
-      state.fever ||
-      state.urgentSymptoms.length > 0 ||
-      state.severePain ||
-      (state.infectionSymptoms.length > 0 && !state.infectionSymptoms.includes('None of the above')) ||
-      (state.recentFall && !state.canBearWeight) ||
-      (state.chronicCondition && state.worseSymptomsToday)
-    );
-  };
-
-  const handleCheckboxChange = (array: string[], item: string, setter: (items: string[]) => void) => {
-    const index = array.indexOf(item);
-    if (index === -1) {
-      setter([...array, item]);
-    } else {
-      setter(array.filter(i => i !== item));
-    }
-  };
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    setSubmitted(true);
-    console.log('Survey submitted:', state);
-  };
-
-
-  const alertStyle = {
-    padding: '15px',
-    backgroundColor: '#ff000020',
-    border: '1px solid #ff0000',
-    borderRadius: '4px',
-    marginBottom: '20px',
-    color: '#cc0000',
-    fontWeight: 'bold',
-  };
 
   return (
-    <div style={containerStyle}>
-      <div style={headerStyle}>
-        <h1 style={{ margin: 0, fontSize: '24px' }}>Medical Assessment Survey</h1>
-      </div>
-
-      <div style={{ ...formContainerStyle, padding: '20px' }}>
-        {/* Emergency Warning Section */}
-        <div style={emergencyWarningStyle}>
-          <div style={warningHeaderStyle}>
-            <span style={{ fontSize: '24px' }}>⚠️</span>
-            <span>EMERGENCY WARNING - CALL 911 IMMEDIATELY IF:</span>
-          </div>
-          <ul style={emergencyListStyle}>
-            <li style={emergencyItemStyle}>
-              <span>•</span>
-              <span>The patient is unconscious, unresponsive, or unable to wake up</span>
-            </li>
-            <li style={emergencyItemStyle}>
-              <span>•</span>
-              <span>The patient is experiencing severe breathing difficulties (gasping, blue lips, unable to speak)</span>
-            </li>
-            <li style={emergencyItemStyle}>
-              <span>•</span>
-              <span>The patient is having severe chest pain that lasts more than a few minutes or radiates to the arm, jaw, or back</span>
-            </li>
-          </ul>
-          <div style={{ marginTop: '15px', fontWeight: 'bold', fontSize: '18px' }}>
-            DO NOT COMPLETE THIS SURVEY - SEEK EMERGENCY CARE
-          </div>
-        </div>
-        <form onSubmit={handleSubmit}>
-          {/* Step 2: Urgent Medical Concerns */}
-          <div style={sectionStyle}>
-            <h2 style={{ color: '#d5002a', marginBottom: '20px' }}>Urgent Medical Concerns</h2>
-            
-            {/* Fever */}
-            <div style={questionStyle}>
-              <label style={labelStyle}>
-                Does the patient have a fever above 102°F (38.9°C)?
-              </label>
-              <div style={radioContainerStyle}>
-                <label>
-                  <input
-                    type="radio"
-                    name="fever"
-                    value="true"
-                    onChange={(e) => setState({ ...state, fever: e.target.value === 'true' })}
-                  /> Yes
-                </label>
-                <label>
-                  <input
-                    type="radio"
-                    name="fever"
-                    value="false"
-                    onChange={(e) => setState({ ...state, fever: e.target.value === 'true' })}
-                  /> No
-                </label>
-              </div>
-            </div>
-
-            {/* Severe pain */}
-            <div style={questionStyle}>
-              <label style={labelStyle}>
-                Is the patient experiencing severe pain anywhere in their body?
-              </label>
-              <div style={radioContainerStyle}>
-                <label>
-                  <input
-                    type="radio"
-                    name="severePain"
-                    value="true"
-                    onChange={(e) => setState({ ...state, severePain: e.target.value === 'true' })}
-                  /> Yes
-                </label>
-                <label>
-                  <input
-                    type="radio"
-                    name="severePain"
-                    value="false"
-                    onChange={(e) => setState({ ...state, severePain: e.target.value === 'true' })}
-                  /> No
-                </label>
-              </div>
-            </div>
-
-            {/* Continue with all other questions using the same pattern... */}
-            {/* Copy the structure from the previous version but apply the new styles */}
-            
-            {/* For checkbox groups, use this structure: */}
-            <div style={questionStyle}>
-              <label style={labelStyle}>
-                Is the patient experiencing any of the following? Check all that apply
-              </label>
-              {['Confusion or disorientation', 'Persistent vomiting',
-                'Severe dehydration (dizziness, dry mouth, no urination)', 'None of the above'].map(symptom => (
-                <div key={symptom} style={checkboxContainerStyle}>
-                  <label>
-                    <input
-                      type="checkbox"
-                      checked={state.urgentSymptoms.includes(symptom)}
-                      onChange={() => handleCheckboxChange(
-                        state.urgentSymptoms,
-                        symptom,
-                        (items) => setState({ ...state, urgentSymptoms: items })
-                      )}
-                    /> {symptom}
-                  </label>
-                </div>
-              ))}
-            </div>
-
-            {/* infectionSymptoms */}
-            <div style={questionStyle}>
-              <label style={labelStyle}>
-                Does the patient have an infection-related symptom? Check all that apply
-              </label>
-              {['Swelling, redness, pus, warmth around a wound', 'Fever with shaking chills',
-                'Increased pain in an existing wound', 'None of the above'].map(symptom => (
-                <div key={symptom} style={checkboxContainerStyle}>
-                  <label>
-                    <input
-                      type="checkbox"
-                      checked={state.infectionSymptoms.includes(symptom)}
-                      onChange={() => handleCheckboxChange(
-                        state.infectionSymptoms,
-                        symptom,
-                        (items) => setState({ ...state, infectionSymptoms: items })
-                      )}
-                    /> {symptom}
-                  </label>
-                </div>
-              ))}
-            </div>
-
-            {/* Continue with all other sections... */}
-          </div>
-
-          {/* Include all other sections with the same structure */}
-          {/* Step 3 and Step 4 sections go here... */}
-        </form>
-      </div>
-
-      <div style={footerStyle}>
-        <button type="submit" style={buttonStyle}>
-          Submit Assessment
+    <div style={{
+      display: 'flex',
+      flexDirection: 'column',
+      alignItems: 'center',
+      justifyContent: 'center',
+      minHeight: '100vh',
+      background: 'linear-gradient(135deg, white 0%, rgba(213, 0, 42, 0.1) 100%)',
+      padding: '20px',
+      fontFamily: '"Century Gothic", CenturyGothic, AppleGothic, sans-serif'
+    }}>
+      <div style={{
+        backgroundColor: 'white',
+        padding: '40px',
+        borderRadius: '16px',
+        maxWidth: '600px',
+        width: '100%',
+        boxShadow: '0 4px 30px rgba(0, 0, 0, 0.05)',
+        textAlign: 'center'
+      }}>
+        <h1 style={{
+          fontSize: '48px',
+          marginBottom: '24px',
+          fontWeight: 'normal',
+          letterSpacing: '2px',
+          background: 'linear-gradient(135deg, #d5002a 0%, #d3058b 100%)',
+          WebkitBackgroundClip: 'text',
+          WebkitTextFillColor: 'transparent',
+          backgroundClip: 'text',
+          color: 'transparent'
+        }}>
+          Healthcare Assistant
+        </h1>
+       
+        <p style={{
+          fontSize: '20px',
+          marginBottom: '40px',
+          lineHeight: '1.6',
+          letterSpacing: '0.5px',
+          color: '#4A4A4A'
+        }}>
+          Welcome to our healthcare portal. Take our health assessment survey to help us understand your needs.
+        </p>
+       
+        {/* Enhanced button with hover effects and shadow */}
+        <button
+          onClick={goToSurvey}
+          onMouseEnter={() => setIsHovered(true)}
+          onMouseLeave={() => setIsHovered(false)}
+          style={{
+            padding: '16px 32px',
+            fontSize: '20px',
+            background: 'linear-gradient(135deg, #d5002a 0%, #d3058b 100%)',
+            color: 'white',
+            border: 'none',
+            borderRadius: '8px',
+            cursor: 'pointer',
+            fontFamily: 'inherit',
+            width: '100%',
+            maxWidth: '300px',
+            marginBottom: '32px',
+            // Dynamic shadow that changes on hover
+            boxShadow: isHovered
+              ? '0 8px 20px rgba(213, 0, 42, 0.3)'
+              : '0 4px 10px rgba(213, 0, 42, 0.2)',
+            // Smooth transitions for hover effects
+            transform: isHovered ? 'translateY(-2px)' : 'translateY(0)',
+            transition: 'all 0.3s ease',
+            // Subtle brightness adjustment on hover
+            filter: isHovered ? 'brightness(1.1)' : 'brightness(1)'
+          }}
+        >
+          Start Health Survey
         </button>
-
-        {submitted && needsUrgentCare() && (
-          <div style={{
-            marginTop: '20px',
-            padding: '15px',
-            backgroundColor: 'rgba(213, 0, 42, 0.1)',
-            border: '1px solid #d5002a',
-            borderRadius: '4px',
+       
+        <div style={{
+          backgroundColor: 'rgba(213, 0, 42, 0.05)',
+          padding: '20px',
+          borderRadius: '8px',
+          maxWidth: '400px',
+          margin: '0 auto'
+        }}>
+          <p style={{
+            fontSize: '16px',
             color: '#d5002a',
-            fontWeight: 'bold',
+            margin: 0,
+            opacity: '0.8'
           }}>
-            URGENT CARE NEEDED: Schedule immediate same-day appointment with high priority.
-          </div>
-        )}
+            If you are experiencing a medical emergency, please call 911 immediately.
+          </p>
+        </div>
       </div>
     </div>
   );
 };
 
-export default MedicalSurvey;
+
+export default HomePage;
